+++ functions/actions/exercises/generateStatutoryConsultationReport.js
@@ -58,8 +58,12 @@ const reportHeaders = (exercise, maxQualificationNum, maxJudicialExperienceNum, 
     ...getQualificationHeaders(maxQualificationNum),
     ...getJudicialExperienceHeaders(exercise, maxJudicialExperienceNum),
     ...getNonJudicialExperienceHeaders(maxNonJudicialExperienceNum),
-    ...getFeePaidOrSalariedHeaders(),
   ];
+
+  if (exercise._applicationVersion < 3) {
+    // only show these fields for application version 2 and below
+    headers.push(...getFeePaidOrSalariedHeaders());
+  }
 
   return headers;
 };
@@ -95,15 +99,21 @@ const reportData = (exercise, applications) => {
     maxJudicialExperienceNum = judicialExperiences.length > maxJudicialExperienceNum ? judicialExperiences.length : maxJudicialExperienceNum;
     maxNonJudicialExperienceNum = nonJudicialExperiences.length > maxNonJudicialExperienceNum ? nonJudicialExperiences.length : maxNonJudicialExperienceNum;
 
-    return {
+    const res = {
       firstName: personalDetails.firstName || null,
       lastName: personalDetails.lastName || null,
       suffix: personalDetails.suffix || null,
       ...getQualificationData(qualifications),
       ...getJudicialExperienceData(exercise, application, judicialExperiences),
       ...getNonJudicialExperienceData(nonJudicialExperiences),
-      ...getFeePaidOrSalariedData(application),
     };
+
+    if (exercise._applicationVersion < 3) {
+      // only show these fields for application version 2 and below
+      res.push(...getFeePaidOrSalariedData(application));
+    }
+
+    return res;
   });
 
   return {
