+++ functions/shared/converters/helpers.js
@@ -146,6 +146,44 @@ const ordinal = (n) => {
   return n + (s[(v - 20) % 10] || s[v] || s[0]);
 };
 
+const getJudicialExperienceString = (exercise, application) => {
+  let judicialExperience = '';
+  if (exercise._applicationVersion >= 3) {
+    let judicialExperiences = [];
+    let quasiJudicialExperiences = [];
+    Array.isArray(application.experience) && application.experience.forEach(experience => {
+      if (experience.jobTitle && experience.judicialFunctions) {
+        if (experience.judicialFunctions.type === 'judicial-post') {
+          judicialExperiences.push(experience.jobTitle);
+        } else if (experience.judicialFunctions.type === 'quasi-judicial-post') {
+          quasiJudicialExperiences.push(experience.jobTitle);
+        }
+      }
+    });
+
+    if (judicialExperiences.length) {
+      judicialExperience += `Judicial - ${judicialExperiences.join(', ')}\n`;
+    }
+    if (quasiJudicialExperiences.length) {
+      judicialExperience += `Quasi-judicial - ${quasiJudicialExperiences.join(', ')}`;
+    }
+
+    if (!judicialExperience) {
+      judicialExperience = `Acquired skills in other way - ${lookup(application.experienceDetails)}`;
+    }
+  } else {
+    if (application.feePaidOrSalariedJudge) {
+      judicialExperience = `Fee paid or salaried judge - ${lookup(application.feePaidOrSalariedSittingDaysDetails)} days`;
+    } else if (application.declaredAppointmentInQuasiJudicialBody) {
+      judicialExperience = `Quasi-judicial body - ${lookup(application.quasiJudicialSittingDaysDetails)} days`;
+    } else {
+      judicialExperience = `Acquired skills in other way - ${lookup(application.skillsAquisitionDetails)}`;
+    }
+  }
+
+  return judicialExperience;
+};
+
 module.exports = {
   addField,
   toYesNo,
@@ -158,4 +196,5 @@ module.exports = {
   attendedUKStateSchool,
   applicationOpenDatePost01042023,
   ordinal,
+  getJudicialExperienceString,
 };
